+++ x/staging/posts/client/cli/keys.go
@@ -21,7 +21,7 @@ const (
 	FlagPollDetails   = "poll-details"
 	FlagPollAnswer    = "poll-answer"
 	FlagCreationTime  = "creation-time"
-	FlagCommentsState = "comments"
+	FlagCommentsState = "comments-state"
 	FlagSubspace      = "subspace"
 	FlagHashtag       = "hashtag"
 	FlagCreator       = "creator"