+++ x/staging/posts/keeper/msgs_server.go
@@ -27,7 +27,7 @@ func computePostID(ctx sdk.Context, msg *types.MsgCreatePost) string {
 		ParentID:             msg.ParentID,
 		Message:              msg.Message,
 		Created:              ctx.BlockTime(),
-		Comments:             msg.Comments,
+		CommentsState:        msg.CommentsState,
 		Subspace:             msg.Subspace,
 		AdditionalAttributes: msg.AdditionalAttributes,
 		Creator:              msg.Creator,
@@ -50,7 +50,7 @@ func (k msgServer) CreatePost(goCtx context.Context, msg *types.MsgCreatePost) (
 		computePostID(ctx, msg),
 		msg.ParentID,
 		msg.Message,
-		msg.Comments,
+		msg.CommentsState,
 		msg.Subspace,
 		msg.AdditionalAttributes,
 		msg.Attachments,
@@ -84,7 +84,7 @@ func (k msgServer) CreatePost(goCtx context.Context, msg *types.MsgCreatePost) (
 				"parent post with id %s not found", post.ParentID)
 		}
 
-		if parentPost.Comments == types.CommentStateBlocked {
+		if parentPost.CommentsState == types.CommentsStateBlocked {
 			return nil, sdkerrors.Wrapf(sdkerrors.ErrInvalidRequest,
 				"post with id %s does not allow comments", parentPost.PostID)
 		}
@@ -139,8 +139,8 @@ func (k msgServer) EditPost(goCtx context.Context, msg *types.MsgEditPost) (*typ
 		existing.PollData = msg.PollData
 	}
 
-	if msg.Comments != types.CommentStateUnspecified {
-		existing.Comments = msg.Comments
+	if msg.CommentsState != types.CommentsStateUnspecified {
+		existing.CommentsState = msg.CommentsState
 	}
 
 	existing.LastEdited = ctx.BlockTime()